move ray refs into daft-partition-refs
