+++ src/daft-distributed/src/python/ray/task.rs
@@ -1,9 +1,9 @@
-use std::{any::Any, collections::HashMap, future::Future, sync::Arc};
+use std::{collections::HashMap, future::Future, sync::Arc};
 
 use common_daft_config::PyDaftExecutionConfig;
 use common_partitioning::{Partition, PartitionRef};
 use daft_local_plan::{ExecutionStats, PyLocalPhysicalPlan, SourceId, python::PyInput};
-use daft_partition_refs::{FlightPartitionRef, PyFlightPartitionRef};
+use daft_partition_refs::{FlightPartitionRef, PyFlightPartitionRef, RayPartitionRef};
 use pyo3::{Py, PyAny, PyResult, Python, pyclass, pymethods};
 
 use crate::{
@@ -139,53 +139,6 @@ impl TaskResultHandle for RayTaskResultHandle {
                 .call_method0(py, "cancel")
                 .expect("Failed to cancel task");
         });
-    }
-}
-
-#[pyclass(module = "daft.daft", name = "RayPartitionRef", frozen, from_py_object)]
-#[derive(Debug, Clone)]
-pub(crate) struct RayPartitionRef {
-    pub object_ref: Arc<Py<PyAny>>,
-    pub num_rows: usize,
-    pub size_bytes: usize,
-}
-
-#[pymethods]
-impl RayPartitionRef {
-    #[new]
-    pub fn new(object_ref: Py<PyAny>, num_rows: usize, size_bytes: usize) -> Self {
-        Self {
-            object_ref: Arc::new(object_ref),
-            num_rows,
-            size_bytes,
-        }
-    }
-
-    #[getter]
-    pub fn get_object_ref(&self, py: Python) -> Py<PyAny> {
-        self.object_ref.clone_ref(py)
-    }
-
-    #[getter]
-    pub fn get_num_rows(&self) -> usize {
-        self.num_rows
-    }
-
-    #[getter]
-    pub fn get_size_bytes(&self) -> usize {
-        self.size_bytes
-    }
-}
-
-impl Partition for RayPartitionRef {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-    fn size_bytes(&self) -> usize {
-        self.size_bytes
-    }
-    fn num_rows(&self) -> usize {
-        self.num_rows
     }
 }
